Refactor UI components and signal connections for more manageable code
Separated views, dialogs, and other UI elements into individual .ui files.
This simplifies maintenance, improves readability, and makes signal connections easier to manage.
