+++ src/assets/about_dialog.ui
@@ -1,0 +1,47 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated with glade 3.38.2 -->
+<interface>
+  <requires lib="gtk+" version="3.24"/>
+  <object class="GtkAboutDialog" id="AboutDialog">
+    <property name="can-focus">False</property>
+    <property name="type-hint">dialog</property>
+    <property name="program-name">Hidamari</property>
+    <property name="copyright" translatable="yes">Copyright © 2022 Jeff Shee</property>
+    <property name="comments" translatable="yes">Video wallpaper for Linux. Written in Python. 🐍
+
+Hidamari 日溜まり 【ひだまり】 (n) sunny spot; exposure to the sun</property>
+    <property name="website">https://github.com/jeffshee/hidamari</property>
+    <property name="website-label" translatable="yes">GitHub</property>
+    <property name="authors">Jeff Shee https://github.com/jeffshee
+All Contributors https://github.com/jeffshee/hidamari/graphs/contributors</property>
+    <property name="artists">Freepik https://www.freepik.com</property>
+    <property name="logo-icon-name">io.github.jeffshee.Hidamari</property>
+    <child internal-child="vbox">
+      <object class="GtkBox">
+        <property name="can-focus">False</property>
+        <property name="orientation">vertical</property>
+        <property name="spacing">2</property>
+        <child internal-child="action_area">
+          <object class="GtkButtonBox">
+            <property name="can-focus">False</property>
+            <property name="layout-style">end</property>
+            <child>
+              <placeholder/>
+            </child>
+            <child>
+              <placeholder/>
+            </child>
+          </object>
+          <packing>
+            <property name="expand">False</property>
+            <property name="fill">False</property>
+            <property name="position">0</property>
+          </packing>
+        </child>
+        <child>
+          <placeholder/>
+        </child>
+      </object>
+    </child>
+  </object>
+</interface>
