+++ src/assets/local_video_view.ui
@@ -1,0 +1,139 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated with glade 3.38.2 -->
+<interface>
+    <requires lib="gtk+" version="3.24"/>
+    <object class="GtkFileFilter" id="filefilter1">
+        <mime-types>
+            <mime-type>text/html</mime-type>
+        </mime-types>
+    </object>
+    <object class="GtkBox" id="LocalVideoView">
+        <property name="visible">True</property>
+        <property name="can-focus">False</property>
+        <property name="orientation">vertical</property>
+        <property name="spacing">8</property>
+        <property name="hexpand">True</property>
+        <property name="vexpand">True</property>
+        <property name="halign">fill</property>
+        <property name="valign">fill</property>
+        <child>
+            <object class="GtkBox">
+            <property name="visible">True</property>
+            <property name="can-focus">False</property>
+            <property name="spacing">8</property>
+            <child>
+                <object class="GtkLabel">
+                <property name="visible">True</property>
+                <property name="can-focus">False</property>
+                <property name="label" translatable="yes">&lt;span size="x-large"&gt;&lt;b&gt;Local Video&lt;/b&gt;&lt;/span&gt;
+Videos located in Hidamari folder</property>
+                <property name="use-markup">True</property>
+                <property name="xalign">0</property>
+                </object>
+                <packing>
+                <property name="expand">True</property>
+                <property name="fill">True</property>
+                <property name="position">0</property>
+                </packing>
+            </child>
+            <child>
+                <object class="GtkButton">
+                <property name="visible">True</property>
+                <property name="can-focus">True</property>
+                <property name="receives-default">True</property>
+                <property name="tooltip-text" translatable="yes">Refresh</property>
+                <property name="halign">center</property>
+                <property name="valign">center</property>
+                <property name="action-name">app.local_video_refresh</property>
+                <property name="always-show-image">True</property>
+                <child>
+                    <object class="GtkImage" id="image2">
+                    <property name="visible">True</property>
+                    <property name="can-focus">False</property>
+                    <property name="icon-name">view-refresh-symbolic</property>
+                    </object>
+                </child>
+                </object>
+                <packing>
+                <property name="expand">False</property>
+                <property name="fill">True</property>
+                <property name="pack-type">end</property>
+                <property name="position">1</property>
+                </packing>
+            </child>
+            <child>
+                <object class="GtkButton">
+                <property name="visible">True</property>
+                <property name="can-focus">True</property>
+                <property name="receives-default">True</property>
+                <property name="tooltip-text" translatable="yes">Open Hidamari folder</property>
+                <property name="halign">center</property>
+                <property name="valign">center</property>
+                <property name="action-name">app.local_video_dir</property>
+                <property name="always-show-image">True</property>
+                <child>
+                    <object class="GtkImage" id="image1">
+                    <property name="visible">True</property>
+                    <property name="can-focus">False</property>
+                    <property name="icon-name">folder-open</property>
+                    </object>
+                </child>
+                </object>
+                <packing>
+                <property name="expand">False</property>
+                <property name="fill">True</property>
+                <property name="pack-type">end</property>
+                <property name="position">2</property>
+                </packing>
+            </child>
+            </object>
+            <packing>
+            <property name="expand">False</property>
+            <property name="fill">True</property>
+            <property name="position">0</property>
+            </packing>
+        </child>
+        <child>
+            <object class="GtkScrolledWindow">
+            <property name="visible">True</property>
+            <property name="can-focus">True</property>
+            <property name="shadow-type">in</property>
+            <child>
+                <object class="GtkIconView" id="IconView">
+                <property name="visible">True</property>
+                <property name="can-focus">True</property>
+                <property name="margin">6</property>
+                <property name="item-width">96</property>
+                <property name="row-spacing">8</property>
+                <property name="column-spacing">8</property>
+                <property name="item-padding">8</property>
+                </object>
+            </child>
+            </object>
+            <packing>
+            <property name="expand">True</property>
+            <property name="fill">True</property>
+            <property name="position">1</property>
+            </packing>
+        </child>
+        <child>
+            <object class="GtkButton" id="ButtonApply">
+            <property name="label">Apply</property>
+            <property name="visible">True</property>
+            <property name="can-focus">True</property>
+            <property name="receives-default">True</property>
+            <property name="halign">end</property>
+            <property name="action-name">app.local_video_apply</property>
+            <style>
+                <class name="suggested-action"/>
+            </style>
+            </object>
+            <packing>
+            <property name="expand">False</property>
+            <property name="fill">True</property>
+            <property name="pack-type">end</property>
+            <property name="position">2</property>
+            </packing>
+        </child>
+        </object>
+</interface>
