+++ src/gui/popover_main.py
@@ -1,0 +1,140 @@
+import sys
+
+try:
+    import os
+    sys.path.insert(1, os.path.join(sys.path[0], ".."))
+    from gui.imports import *
+    from gui.about_dialog import AboutDialog
+    from utils import ConfigUtil
+except ModuleNotFoundError:
+    from hidamari.gui.imports import *
+    from hidamari.gui.about_dialog import AboutDialog
+    from hidamari.utils import ConfigUtil
+
+class PopoverMain:
+    def __init__(self, config: ConfigUtil, server):
+        self.config = config
+        self.server = server
+        self.builder = Gtk.Builder()
+        try:
+            self.builder.add_from_resource(APP_UI_RESOURCE_PATH + "popover_main.ui")
+        except GLib.Error:
+            self.builder.add_from_file(os.path.abspath("./assets/popover_main.ui"))
+
+        self.widget = self.builder.get_object("PopoverMain")
+
+        signals = {
+            "on_volume_changed": self.on_volume_changed,
+            "on_blur_radius_changed": self.on_blur_radius_changed,
+        }
+        
+        self.builder.connect_signals(signals)
+
+    def on_volume_changed(self, adjustment):
+        self.config[CONFIG_KEY_VOLUME] = int(adjustment.get_value())
+        logger.info(f"[GUI/PopoverMain] Volume: {self.config[CONFIG_KEY_VOLUME]}")
+        self._save_config_delay()
+        if self.server is not None:
+            self.server.volume = self.config[CONFIG_KEY_VOLUME]
+        self.set_mute_toggle_icon()
+
+    def on_blur_radius_changed(self, adjustment):
+        self.config[CONFIG_KEY_BLUR_RADIUS] = int(adjustment.get_value())
+        logger.info(f"[GUI/PopoverMain] Blur radius: {self.config[CONFIG_KEY_BLUR_RADIUS]}")
+        self._save_config_delay()
+        if self.server is not None:
+            self.server.blur_radius = self.config[CONFIG_KEY_BLUR_RADIUS]
+        
+    def on_popover_main_activate(self):
+        print("PopoverMain activated")
+        
+    def set_mute_toggle_icon(self):
+        toggle_icon: Gtk.Image = self.builder.get_object("ToggleMuteIcon")
+        volume, is_mute = self.config[CONFIG_KEY_VOLUME], self.config[CONFIG_KEY_MUTE]
+        if volume == 0 or is_mute:
+            icon_name = "audio-volume-muted-symbolic"
+        elif volume < 30:
+            icon_name = "audio-volume-low-symbolic"
+        elif volume < 60:
+            icon_name = "audio-volume-medium-symbolic"
+        else:
+            icon_name = "audio-volume-high-symbolic"
+        toggle_icon.set_from_icon_name(icon_name=icon_name, size=0)
+
+    def set_scale_volume_sensitive(self):
+        scale = self.builder.get_object("ScaleVolume")
+        if self.config[CONFIG_KEY_MUTE]:
+            scale.set_sensitive(False)
+        else:
+            scale.set_sensitive(True)
+
+    def set_spin_blur_radius_sensitive(self):
+        spin = self.builder.get_object("SpinBlurRadius")
+        if self.config[CONFIG_KEY_STATIC_WALLPAPER]:
+            spin.set_sensitive(True)
+        else:
+            spin.set_sensitive(False)
+
+    def on_mute(self, action, state):
+        action.set_state(state)
+        self.config[CONFIG_KEY_MUTE] = bool(state)
+        logger.info(f"[GUI/PopoverMain] {action.get_name()}: {state}")
+        self._save_config()
+        if self.server is not None:
+            self.server.is_mute = self.config[CONFIG_KEY_MUTE]
+        self.set_mute_toggle_icon()
+        self.set_scale_volume_sensitive()
+        
+    def on_static_wallpaper(self, action, state):
+        action.set_state(state)
+        self.config[CONFIG_KEY_STATIC_WALLPAPER] = bool(state)
+        logger.info(f"[GUI/PopoverMain] {action.get_name()}: {state}")
+        self._save_config()
+        if self.server is not None:
+            self.server.is_static_wallpaper = self.config[CONFIG_KEY_STATIC_WALLPAPER]
+        self.set_spin_blur_radius_sensitive()
+
+    def on_pause_when_maximized(self, action, state):
+        action.set_state(state)
+        self.config[CONFIG_KEY_PAUSE_WHEN_MAXIMIZED] = bool(state)
+        logger.info(f"[GUI/PopoverMain] {action.get_name()}: {state}")
+        self._save_config()
+        if self.server is not None:
+            self.server.is_pause_when_maximized = self.config[CONFIG_KEY_PAUSE_WHEN_MAXIMIZED]
+        
+    def on_mute_when_maximized(self, action, state):
+        action.set_state(state)
+        self.config[CONFIG_KEY_MUTE_WHEN_MAXIMIZED] = bool(state)
+        logger.info(f"[GUI/PopoverMain] {action.get_name()}: {state}")
+        self._save_config()
+        if self.server is not None:
+            self.server.is_mute_when_maximized = self.config[CONFIG_KEY_MUTE_WHEN_MAXIMIZED]
+            
+    def on_about(self, window: Gtk.Window, version: str, *_):
+        dialog = AboutDialog().dialog
+        dialog.set_transient_for(window)
+        dialog.set_version(version)
+        dialog.present()
+        
+    def set_mute_toggle_icon(self):
+        toggle_icon: Gtk.Image = self.builder.get_object("ToggleMuteIcon")
+        volume, is_mute = self.config[CONFIG_KEY_VOLUME], self.config[CONFIG_KEY_MUTE]
+        if volume == 0 or is_mute:
+            icon_name = "audio-volume-muted-symbolic"
+        elif volume < 30:
+            icon_name = "audio-volume-low-symbolic"
+        elif volume < 60:
+            icon_name = "audio-volume-medium-symbolic"
+        else:
+            icon_name = "audio-volume-high-symbolic"
+        toggle_icon.set_from_icon_name(icon_name=icon_name, size=0)
+        
+    def _load_config(self):
+        self.config = ConfigUtil().load()
+        
+    def _save_config(self):
+        ConfigUtil().save(self.config)
+        
+    @debounce(1)
+    def _save_config_delay(self):
+        self._save_config()
